requirements.txt
cython
scipy
shapely
timm
scikit-image
numpy==1.23.1
setuptools==59.5.0
matplotlib==3.3.4
pillow==9.1.0
shapely==1.8.0
open-clip-torch
einops
resampy
soundfile
easydict